fix: mobile carousel swiping improvements
